% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{H3K27.meta}
\alias{H3K27.meta}
\title{Annotation for the H3K27 e expression matrix for scTHI.}
\format{An object of class \code{data.frame} with 527 rows and 9 columns.}
\usage{
H3K27.meta
}
\description{
A dataframe 527x9 for the  matrix of patient PatientBCH836 form Filbin,
M. G., Tirosh, I.,
Hovestadt, V., Shaw, M. L.,
Escalante, L. E., Mathewson, N. D., ... & Haberler, C. (2018).
Developmentaland oncogenic programs in H3K27M
gliomas dissected by single-cell RNA-seq.
Science, 360(6386), 331-335.
H3K27.annotation
}
\keyword{datasets}
