% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{H3K27}
\alias{H3K27}
\title{Example expression matrix for scTHI.}
\format{An object of class \code{matrix} with 21673 rows and 527 columns.}
\usage{
H3K27
}
\description{
A 21673 x 5527 matrix patient PatientBCH836 form Filbin,
M. G., Tirosh, I.,
Hovestadt, V., Shaw, M. L.,
Escalante, L. E., Mathewson, N. D., ... & Haberler, C. (2018).
Developmentaland oncogenic programs in H3K27M
gliomas dissected by single-cell RNA-seq.
Science, 360(6386), 331-335.
H3K27
}
\keyword{datasets}
