% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MGH45.annotation}
\alias{MGH45.annotation}
\title{Annotation for Example expression matrix for scTHI.}
\format{An object of class \code{data.frame} with 608 rows and 2 columns.}
\usage{
MGH45.annotation
}
\description{
A 608 x 2 dataframe with the annotation for the MGH4 matrix
Venteicher AS, Tirosh I, Hebert C, Yizhak K et al.
Decoupling genetics, lineages, and microenvironment
in IDH-mutant gliomas by single-cell RNA-seq.
Science 2017 Mar 31;355(6332)
MGH45.annotation
}
\keyword{datasets}
