% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MGH45}
\alias{MGH45}
\title{Example expression matrix for scTHI.}
\format{An object of class \code{matrix} with 17584 rows and 608 columns.}
\usage{
MGH45
}
\description{
A 17584  x 608 matrix patient PatientBCH836 form
Venteicher AS, Tirosh I, Hebert C, Yizhak K et al.
Decoupling genetics, lineages, and microenvironment
in IDH-mutant gliomas by single-cell RNA-seq.
Science 2017 Mar 31;355(6332)
MGH45
}
\keyword{datasets}
